cassandra/requirements.txt
azure-functions
Cassandra-driver 
prettytable
requests
pyopenssl